feat(DQN): Move other PG models to a common directory
