+++ rlberry/agents/cem/cem.py
@@ -9,20 +9,9 @@ import rlberry.spaces as spaces
 from rlberry.agents import Agent
 
 # choose device
+from rlberry.agents.common.models import Net
+
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-
-
-class Net(nn.Module):
-    def __init__(self, obs_size, hidden_size, n_actions):
-        super(Net, self).__init__()
-        self.net = nn.Sequential(
-            nn.Linear(obs_size, hidden_size),
-            nn.ReLU(),
-            nn.Linear(hidden_size, n_actions)
-        )
-
-    def forward(self, x):
-        return self.net(x)
 
 
 class Memory:
